+++ CHANGELOG.md
@@ -8,3 +8,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ### Added
  - A changelog.
+
+## 2025-09-30
+ 
+### Removed
+ - All jsDelivr related files.